Rename appRoot to Root and devTools to DevTools
