+++ app/redux/containers/Root.dev.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { Provider } from 'react-redux';
 import { Router, browserHistory } from 'react-router';
 
-import DevTools from './devTools';
+import DevTools from './DevTools';
 
 export default class Root extends Component {
   render() {
